Restructuring newsfeed-related folders
