+++ Antrashare_ui/src/app/app-routing.module.ts
@@ -6,7 +6,7 @@ import { MyProfilePageComponent } from './components/my-profile-page/my-profile-
 
 import { SettingsComponent } from './components/settings/settings.component';
 import { NewsFeedComponent } from './components/news-feed/news-feed.component';
-import { StoryComponent } from './components/story/story.component';
+import { StoryComponent } from './components/news-feed/story/story.component';
 import { ErrorPageComponent } from './components/error-page/error-page.component';
 
 
@@ -19,7 +19,7 @@ const routes: Routes = [
   {path:'settings',component:SettingsComponent},
   {path:'story',component:StoryComponent},
   { path: '**', component: ErrorPageComponent }
- 
+
 ];
 
 @NgModule({
