+++ Antrashare_ui/src/app/components/news-feed/story/story.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles.scss';
+@import '../../../../styles.scss';
 
 .story {
     border: $red solid 1px;
@@ -8,22 +8,31 @@
 }
 .storyHeader {
     display: grid;
-    grid-template-columns: 10% 20% 70%;
-    padding: 2% 5%;
+    grid-template-columns: 15% 40% auto;
+    padding: 4% 5%;
     color: $black;
+    background-color: $white;
+    margin: 2% 5%;
 }
 
-h2 {
-    font-size: 1.5rem;
+.storyHeader mat-card-title {
+    padding-bottom: 5%;
 }
 
 #publicationInfo {
-    display: grid;
-    grid-template-rows: 90% 10%;
-    height: 100%;
+    height: 80%;
+}
+
+.avatar-card {
+    height: 80%;
+    width: 80%;
+    margin: 0 auto;
+    text-align: center;
 }
 
 .storyContent {
+    display: grid;
+    grid-template-rows: 50% 50%;
     height: 30rem;
     margin: 2% 5%;
     border-radius: 4px;
@@ -32,19 +41,29 @@ h2 {
     overflow: auto;
 }
 
+.storyMedia {
+    display: grid;
+    grid-template-columns: 50% 50%;
+    align-content: center;
+    height: 100%;
+    width: 100%;
+    padding: 0 auto;
+    margin: 0 auto;
+}
+
 .storyBottomBar {
     display: grid;
     grid-template-columns: 33% 34% 33%;
     text-align: center;
     margin: 2% 5%;
     height: 100%;
+    background-color: $white;
 }
 
 .storyBottomBarButton {
     background-color: $red;
     color: $white;
     transition: 0.3s ease-in-out;
-    // margin-bottom: 5%;
 }
 
 .storyBottomBarButton:hover {
