+++ Antrashare_ui/src/app/components/news-feed/story/story.component.ts
@@ -1,3 +1,4 @@
+import { TOUCH_BUFFER_MS } from '@angular/cdk/a11y/input-modality/input-modality-detector';
 import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
@@ -11,34 +12,37 @@ import { News } from 'src/app/models/newsfeed.models';
 export class StoryComponent implements OnInit {
   @Input() story!:News
   @Output() storyEmiter = new EventEmitter()
+
   public isCommentOpened = false;
   public hasVideo = false;
   public hasImage = false;
+  public noVideo = true;
+  public noImage = true;
   videoLink: any;
   imageLink: any;
   safeVideo: any;
   safeImage: any;
 
-  constructor(private _sanitizer: DomSanitizer){
-    this.safeVideo = this._sanitizer.bypassSecurityTrustResourceUrl(this.videoLink);
-    this.safeImage = this._sanitizer.bypassSecurityTrustResourceUrl(this.imageLink);
-    if (this.safeVideo) {
+  constructor(private _sanitizer: DomSanitizer) { }
+
+  ngOnInit(): void {
+    this.videoLink = this.story.content.video ? this.story.content.video : '';
+    this.imageLink = this.story.content.image ? this.story.content.image : '';
+
+    if (this.videoLink !== '') {
       this.hasVideo = true;
     } else {
       this.hasVideo = false;
     }
-    if (this.safeImage) {
+
+    if (this.imageLink !== '') {
       this.hasImage = true;
     } else {
       this.hasImage = false;
     }
- }
 
-  ngOnInit(): void {
-    if (this.story.content.video) {
-      this.videoLink = this.story.content.video;
-    }
-    console.log(this.story.content.image);
+    this.safeVideo = this._sanitizer.bypassSecurityTrustResourceUrl(this.videoLink);
+    this.safeImage = this._sanitizer.bypassSecurityTrustResourceUrl(this.imageLink);
   }
 
   onTriggerComment() {
