--- Antrashare_ui/src/app/components/story/story.component.html
@@ -1,36 +1,0 @@
-<div class="story">
-    <div class="storyHeader">
-        <div>
-            Avatar
-        </div>
-        <div id='publicationInfo'>
-            <h2>
-                {{story.publisherName}}
-            </h2>
-            <h3>
-                Posted on: {{story.publishedTime}}
-            </h3>
-        </div>
-    </div>
-    <div class="storyContent">
-        {{story.content.text}} <br/>
-        <img *ngIf='hasImage' [src]='safeImage'><br/>
-        <iframe *ngIf="hasVideo" [src]='safeVideo' frameborder="0" allowfullscreen></iframe>
-    </div>
-    <div class="storyBottomBar">
-        <div class="storyBottomBarItem">
-            <button mat-raised-button class="storyBottomBarButton" (click)="onTriggerComment()">
-                <mat-icon>comment</mat-icon>
-            </button>
-        </div>
-        <!-- <div class="storyBottomBarItem">
-            Pagination
-        </div> -->
-        <div class="storyBottomBarItem">
-            <button mat-raised-button class="storyBottomBarButton">
-                <mat-icon>thumb_up</mat-icon>
-            </button>
-        </div>
-    </div>
-</div>
-<app-story-comments *ngIf="isCommentOpened"></app-story-comments>
